sc - attempt to move App into a folder
