+++ frontend/src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./app/App";
 import Auth0ProviderWithHistory from "components/Auth/Auth0ProviderWithHistory";
 import { BrowserRouter } from "react-router-dom";
 import * as serviceWorker from "./serviceWorker";
